services/job_service/backend/requirements.txt
libmagic
python-multipart
pytesseract
pdf2image
python-dotenv
pillow
python-magic
scikit-learn
nltk
fastapi
fastapi[all]
uvicorn
flask 
flask-sqlalchemy
pdfminer.six==20221105
python-docx==0.8.11
aio_pika
PyMuPDF 
pillow 
pdfminer.six 
python-docx 
requests
httpx

tenacity
aiofiles
pymongo

python-multipart
python-dotenv

# AI and Processing
google-generativeai
PyPDF2
docx2txt
pytesseract
pdf2image
config
# Development
pytest
pytest-asyncio
httpx


boto3